+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v1.1.37 (1.20.1)
 - fix client issue for null-attribute
 - added config option to toggle lore injection on server
+- minor ui-optimisations