+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## v1.1.39 (1.20.1)
 - added handheld Epic Fight Compat
+- added config option to toggle UI scaling