+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## v1.1.39 (1.20.1)
+## v1.1.38 (1.20.1)
+- added faster mining modifier
+- adjusted default enchanting glint color
+- minor changes to disable glint shader registration as well if glint is disabled in the config
 - added handheld Epic Fight Compat
 - added config option to toggle UI scaling